backend/requirements.txt
aniso8601==6.0.0
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.3
python-dotenv==0.21.0
pytz==2019.1
six==1.12.0
SQLAlchemy==1.4.40
Werkzeug==2.2.2
zipp==3.8.1
